changes in FPSHandler (uint->double), slash stuff

+++ civ/FPSHandler.cpp
@@ -1,7 +1,7 @@
 #include "FPSHandler.h"
 
 FPSHandler::FPSHandler(uint32_t desiredFPS_):
-	desiredFPS{ desiredFPS_ },
+	desiredFPS{ static_cast<double>(desiredFPS_) },
 	delay{ 25 }{
 
 }
@@ -15,9 +15,9 @@ void FPSHandler::Align(uint32_t ticks){
 	lastFPS = ticks / diff.count();
 
 	if (lastFPS > 1.05 * desiredFPS)
-		delay *= 1.1;
+		delay *= 1.1f;
 	else if (lastFPS < 0.95 * desiredFPS)
-		delay *= 0.9;
+		delay *= 0.9f;
 
 	lastTime = now;
 }
